Feat: Add OptimizeHtml Listener; Add optimizers property to config
