+++ config/page-cache.php
@@ -9,4 +9,9 @@ return [
     'cache_pagination' => false,
 
     'filesystem' => 'page-cache',
+
+    'optimizers' => [
+        \SiteOrigin\PageCache\Jobs\CriticalCssJob::class,
+        \SiteOrigin\PageCache\Jobs\HTMLMinifierJob::class,
+    ]
 ];
